linter rules applied & fixed imports and minor bugs

+++ lib/widgets/sliders/volume_slider.dart
@@ -21,7 +21,7 @@ class VolumeSlider extends StatelessWidget {
         SoundManager.instance.setVolume(value);
         await AppServices.instance.localStorageService.setIntValue(
           LocalStorageKey.volume, 
-          value.ceil().toInt()
+          value.ceil(),
         );
       },
       appearance: CircularSliderAppearance(
